+++ fahrschein-e2e-test/pom.xml
@@ -41,6 +41,12 @@
         </dependency>
         <dependency>
             <groupId>org.zalando</groupId>
+            <artifactId>fahrschein-http-simple</artifactId>
+            <version>${project.version}</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.zalando</groupId>
             <artifactId>fahrschein-http-apache</artifactId>
             <version>${project.version}</version>
             <scope>test</scope>
