+++ fahrschein-test-support/pom.xml
@@ -53,6 +53,10 @@
             <artifactId>junit</artifactId>
             <version>${version.junit}</version>
         </dependency>
-
+        <dependency>
+            <groupId>org.mockito</groupId>
+            <artifactId>mockito-core</artifactId>
+            <version>${version.mockito}</version>
+        </dependency>
     </dependencies>
 </project>
